feat(arch): PR 5/12 — move credential_store.py to infrastructure/stores
- Move credential_store.py from core/stores to infrastructure/stores
- Update imports in cli_bot.py, cli_setup.py, bootstrap_stores.py,
  bootstrap_wiring.py, adapter_standalone.py
- Update test imports in test_credential_store.py, test_bot_commands.py,
  test_cli_setup.py
- Remove credential_store transitional ignore from .importlinter
- Update infrastructure/CLAUDE.md to mark credential_store as migrated
