+++ src/lyra/bootstrap/adapter_standalone.py
@@ -12,7 +12,7 @@ from lyra.adapters.nats_outbound_listener import NatsOutboundListener
 from lyra.bootstrap.signal_handlers import setup_shutdown_event
 from lyra.core.bus import Bus
 from lyra.core.message import InboundMessage, Platform
-from lyra.core.stores.credential_store import CredentialStore, LyraKeyring
+from lyra.infrastructure.stores.credential_store import CredentialStore, LyraKeyring
 from lyra.nats.queue_groups import adapter_outbound
 from roxabi_nats import nats_connect
 from roxabi_nats.connect import scrub_nats_url
