+++ src/lyra/infrastructure/stores/credential_store.py
