+++ tests/cli/test_bot_commands.py
@@ -21,7 +21,7 @@ import pytest
 from typer.testing import CliRunner
 
 from lyra.cli_bot import bot_app
-from lyra.core.stores.credential_store import CredentialStore, LyraKeyring
+from lyra.infrastructure.stores.credential_store import CredentialStore, LyraKeyring
 
 # ---------------------------------------------------------------------------
 # Helpers
